pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sys-info-api"
version = "0.0.1"
authors = [
    { name = "Charlie Powell", email = "cdp1337@veraciousnetwork.com" },
]
description = "Collection of APIs to retrieve various system information for Linux, BSD, and MacOS."
readme = "README.md"
license = {text = "GNU AGPLv3"}
requires-python = ">=3.8"
classifiers = [
    # See https://pypi.org/classifiers/ for full list
    "Development Status :: 1 - Planning",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: BSD",
    "Operating System :: POSIX :: Linux",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Topic :: System :: Filesystems",
    "Topic :: System :: Hardware",
    "Topic :: System :: Networking",
    "Topic :: System :: Operating System",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "python-dateutil>=2.8.2",
    "six>=1.16.0",
    "ipcalc>=1.99.0",
    "requests>=2.31.0",
    "pyyaml>=5.4.1",
]
# keywords = ["egg", "bacon", "sausage", "tomatoes", "Lobster Thermidor"]

[project.optional-dependencies]
dev = [
    "coverage>=5.0.3",
    "flake8>=3.9.2",
    "pytest",
    "pytest-benchmark[histogram]>=3.2.1",
]

[project.urls]
Homepage = "https://github.com/cdp1337/sys-info-api"
Issues = "https://github.com/cdp1337/sys-info-api/issues"

[project.scripts]
generate_test_data = "sys_info_api.bin.generate_test_data:run"